% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copykat.R
\name{Copykat}
\alias{Copykat}
\title{Spatial transcriptomic cancer spot prediction}
\usage{
Copykat(
  obj,
  cancer = "cancer",
  n_PC = 10,
  id.type = "S",
  cell.line = "no",
  ngene.chr = 5,
  LOW.DR = 0.05,
  UP.DR = 0.1,
  win.size = 25,
  norm.cell.names = "",
  KS.cut = 0.1,
  sam.name = "",
  distance = "euclidean",
  output.seg = "FALSE",
  plot.genes = "TRUE",
  genome = "hg20",
  n.cores = 1
)
}
\arguments{
\item{obj}{Spatial transcriptome object.}

\item{cancer}{Cancer type.}

\item{n_PC}{Dimensions of reduction to use as input.}

\item{id.type}{Gene id type: Symbol or Ensemble.}

\item{cell.line}{If the data are from pure cell line,put "yes"; if cell line data are a mixture of tumor and normal cells, still put "no".}

\item{ngene.chr}{Minimal number of genes per chromosome for cell filtering.}

\item{LOW.DR}{Minimal population fractions of genes for smoothing.}

\item{UP.DR}{Minimal population fractions of genes for segmentation.}

\item{win.size}{Minimal window sizes for segmentation.}

\item{norm.cell.names}{A vector of normal cell names.}

\item{KS.cut}{Segmentation parameters, input 0 to 1; larger looser criteria.}

\item{sam.name}{Sample name.}

\item{distance}{Distance methods include euclidean, and correlation converted distance include pearson and spearman.}

\item{output.seg}{TRUE or FALSE, output seg file for IGV visualization.}

\item{plot.genes}{TRUE or FALSE, output heatmap of CNVs with genename labels.}

\item{genome}{hg20 or mm10, current version only work for human or mouse genes.}

\item{n.cores}{Number of cores for parallel computing.}
}
\value{
A list. containing predicted object, cancer object and normal object.
}
\description{
Use copykat to predict cancer spots and split normal spots and cancer spots.
}
\examples{
\dontrun{test <- Copykat(obj,cancer = "colon cancer",n_PC = 10,id.type="S", ngene.chr=5, win.size=25, KS.cut=0.1, sam.name="GSM7058756_C1", distance="euclidean", norm.cell.names="")}
}
