% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSC.R
\name{SSC}
\alias{SSC}
\title{Spatial spot-spot communication}
\usage{
SSC(obj, json_path, spot.size = 65)
}
\arguments{
\item{obj}{Object containing the cell type.}

\item{json_path}{The path of json.}

\item{spot.size}{The theoretical spot size (um) in 10X Visium}
}
\value{
The result of spot-spot communication.
}
\description{
Using cellchat to calculate spatial transcriptome spot-spot communication network.
}
\examples{
\dontrun{ssc_pre <- SSC(obj,json_path = "../data/spatial/scalefactors_json.json")}
}
