% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSEA.R
\name{SSEA}
\alias{SSEA}
\title{Spatial Gene Sets Enrichment Analysis}
\usage{
SSEA(obj_list, num_list, cancer = cancer, population_size, method = "BH")
}
\arguments{
\item{obj_list}{A seurat object list. Includes whole spatial transcriptome data and data split by disease and normal controls.}

\item{num_list}{A numeric list. e.g. seq(100,1000,100).}

\item{cancer}{Cancer type.}

\item{population_size}{Number of genes in the species studied.}

\item{method}{Statistical methods for p-value correction. c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")}
}
\value{
A list, containing spot and cluster cell type prediction and predicted object.
}
\description{
Separate annotation for tumor spots and normal spots processed data. Integrate tumor and normal control predictions into original object.
}
\examples{
\dontrun{pred <- SSEA(obj_list,num_list,"colon cancer",20000)}
}
