% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSEA_score_normal.R
\name{SSEA_score_normal}
\alias{SSEA_score_normal}
\title{Spatial genesets enrichment score for normal spots}
\usage{
SSEA_score_normal(HallMakrer, test_gene, population_size, method = "BH")
}
\arguments{
\item{HallMakrer}{Gene markers for normal spots.}

\item{test_gene}{Top_gene lists.}

\item{population_size}{Number of genes in the species studied.}

\item{method}{Statistical methods for p-value correction. c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none")}
}
\value{
A cell type prediction list, containing spot and cluster cell type prediction.
}
\description{
Predicting cell types in normal regions from spatial transcriptomic data.
}
\examples{
\dontrun{pred <- SSEA_score_normal(HallMakrer,test_gene,20000)}
}
