% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_gene.R
\name{top_gene}
\alias{top_gene}
\title{Top n genes}
\usage{
top_gene(mat, n)
}
\arguments{
\item{mat}{A gene expression matrix. The row name is the gene name and the column name is spot.}

\item{n}{A numeric list. e.g. seq(100,1000,100).}
}
\value{
The default method returns a data frame list.
}
\description{
According to the provided numeric list, stat the genes whose expression level is in the top n for each spot.
}
\examples{
\dontrun{top_gene(mat,num_list)}
}
